requirements-dev.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-bootstrap-form==3.4
djangorestframework==3.12.2
drf-yasg==1.20.0
future==0.18.2
html5lib==1.1
httpie==2.3.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
packaging==20.8
Pillow==8.0.1
Pygments==2.7.4
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
PySocks==1.7.1
python-bidi==0.4.2
pytz==2020.4
PyYAML==5.4.1
reportlab==3.5.59
requests==2.25.1
requests-toolbelt==0.9.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.2
webencodings==0.5.1
xhtml2pdf==0.2.5
xlwt==1.3.0
